MachinLearningMicroservices/mds_image/requirements.txt
numpy
flask
configparser
Pillow
requests
pywebhdfs
